+++ src/code/const_init.rs
@@ -164,20 +164,30 @@ pub use _impl_init;
 mod impl_core {
     use super::ConstInit;
     use crate::{
+        // cell
+        Cell, OnceCell, RefCell, UnsafeCell,
+        // cmp
+        Reverse, Ordering,
+        // ffi
+        CStr,
+        // marker
+        PhantomData, PhantomPinned,
+        // mem
+        ManuallyDrop,
+        // num
+        FloatCategory, Saturating, Wrapping,
         NonZeroU8, NonZeroU16, NonZeroU32, NonZeroU64, NonZeroU128, NonZeroUsize,
         NonZeroI8, NonZeroI16, NonZeroI32, NonZeroI64, NonZeroI128, NonZeroIsize,
-        Cell, OnceCell, RefCell, UnsafeCell,
-        PhantomData, PhantomPinned, ManuallyDrop,
-        Reverse, Saturating, Wrapping,
+        // ops
+        Bound, ControlFlow,
         Range, RangeFrom, RangeFull, RangeInclusive, RangeTo, RangeToInclusive,
+        // panic
         PanicAssertUnwindSafe,
+        // range WAIT:1.96
+        // RangeInclusive,
+        // time
+        Duration,
     };
-    // TEMP during merge:
-    use ::core::time::Duration;
-    use ::core::ffi::CStr;
-
-    // TODO: Types that don't implement `Default`:
-    // ops::{Bound, ControlFlow, CoroutineState, FpCategory, Ordering, Result},
     // NOTE: atomic types are implemented in devela::work::sync::atomic
 
     /* tuples, arrays */
@@ -187,16 +197,42 @@ mod impl_core {
         0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12,
         13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 32];
 
-    /* non-generic */
-
+    // *primitives*
     _impl_init![ConstInit: () => ()];
     _impl_init![ConstInit: false => bool];
     _impl_init![ConstInit: '\x00' => char];
-
     _impl_init![ConstInit: 0 =>
         i8, i16, i32, i64, i128, isize, u8, u16, u32, u64, u128, usize];
     _impl_init![ConstInit: 0.0 => f32, f64];
+    _impl_init![ConstInit: <T> core::ptr::null() => *const T];
+    _impl_init![ConstInit: <T> core::ptr::null_mut() => *mut T];
+    _impl_init![ConstInit: <T> &[] => &[T]]; // not allowed for &mut [T]
+    _impl_init![ConstInit: "" => &str];
+    #[cfg(all(not(feature = "safe_text"), feature = "unsafe_str"))]
+    #[cfg_attr(nightly_doc, doc(cfg(feature = "unsafe_str")))]
+    impl crate::ConstInit for &mut str {
+        // SAFETY: The empty string is valid UTF-8.
+        const INIT: Self = unsafe { ::core::str::from_utf8_unchecked_mut(&mut []) }; }
 
+    // cell
+    _impl_init![ConstInit: <T: ConstInit> Self::new() => OnceCell<T>];
+    _impl_init![ConstInit: <T: ConstInit> Self::new(T::INIT) =>
+        Cell<T>, ManuallyDrop<T>, RefCell<T>, UnsafeCell<T>];
+
+    // cmp
+    _impl_init![ConstInit: Ordering::Equal => Ordering];
+    _impl_init![ConstInit: <T: ConstInit> Self(T::INIT) => Reverse<T>];
+
+    // ffi
+    impl ConstInit for &CStr { const INIT: Self = {
+        if let Ok(s) = CStr::from_bytes_until_nul(&[0]) { s } else { unreachable![]; } };
+    }
+
+    // marker
+    _impl_init![ConstInit: <T> Self => PhantomData<T>]; // no need for trait bound here
+    _impl_init![ConstInit: Self => PhantomPinned];
+
+    // num
     crate::CONST! { _NZ = "This implementation returns the same as [`Self::MIN`]."; }
     _impl_init![ConstInit: Self::MIN =>
         #[doc=_NZ!()]NonZeroU8,  #[doc=_NZ!()]NonZeroU16,  #[doc=_NZ!()]NonZeroU32,
@@ -204,54 +240,33 @@ mod impl_core {
         #[doc=_NZ!()]NonZeroI8,  #[doc=_NZ!()]NonZeroI16,  #[doc=_NZ!()]NonZeroI32,
         #[doc=_NZ!()]NonZeroI64, #[doc=_NZ!()]NonZeroI128, #[doc=_NZ!()]NonZeroIsize
     ];
-    _impl_init![ConstInit: Duration::new(0, 0) => Duration];
+    _impl_init![ConstInit: <T: ConstInit> Self(T::INIT) => Saturating<T>, Wrapping<T>];
+    _impl_init![ConstInit: FloatCategory::Nan => FloatCategory];
 
+    // ops
+    _impl_init![ConstInit: <T> Self::Unbounded => Bound<T> ];
+    _impl_init![ConstInit: <B: ConstInit, C: ConstInit>
+        Self::Continue(C::INIT) => ControlFlow<B, C> ];
+    _impl_init![ConstInit: <T: ConstInit> Self { start: T::INIT, end: T::INIT } => Range<T>];
+    _impl_init![ConstInit: <T: ConstInit> Self { start: T::INIT } => RangeFrom<T>];
     _impl_init![ConstInit: Self => RangeFull];
-    _impl_init![ConstInit: Self => PhantomPinned];
+    _impl_init![ConstInit: <T: ConstInit> Self::new(T::INIT, T::INIT) => RangeInclusive<T>];
+    _impl_init![ConstInit: <T: ConstInit> Self { end: T::INIT } => RangeTo<T>, RangeToInclusive<T>];
+    #[cfg(nightly_coro)]
+    _impl_init![ConstInit: <Y: ConstInit, R: ConstInit>
+        Self::Yielded(Y::INIT) => crate::CoroutineState<Y, R> ];
 
-    impl ConstInit for &CStr {
-        const INIT: Self = {
-            if let Ok(s) = CStr::from_bytes_until_nul(&[0]) { s } else { unreachable![]; } }; }
-    _impl_init![ConstInit: "" => &str];
-
-    #[cfg(all(not(feature = "safe_text"), feature = "unsafe_str"))]
-    #[cfg_attr(nightly_doc, doc(cfg(feature = "unsafe_str")))]
-    impl crate::ConstInit for &mut str {
-        // SAFETY: The empty string is valid UTF-8.
-        const INIT: Self = unsafe { ::core::str::from_utf8_unchecked_mut(&mut []) };
-    }
-
-    /* generic unbounded */
-
-    _impl_init![ConstInit: <T> core::ptr::null() => *const T];
-    _impl_init![ConstInit: <T> core::ptr::null_mut() => *mut T];
-    _impl_init![ConstInit: <T> &[] => &[T]]; // not allowed for &mut [T]
-
-    /* generic bounded */
-
-    _impl_init![ConstInit: <T: ConstInit>
-        Self { start: T::INIT } => RangeFrom<T>];
-    _impl_init![ConstInit: <T: ConstInit>
-        Self { end: T::INIT } => RangeTo<T>, RangeToInclusive<T>];
-    _impl_init![ConstInit: <T: ConstInit>
-        Self { start: T::INIT, end: T::INIT } => Range<T>];
-    _impl_init![ConstInit: <T: ConstInit>
-        Self::new(T::INIT, T::INIT) => RangeInclusive<T>]; // this one has private fields
-
-    _impl_init![ConstInit: <T: ConstInit> Self::new() => OnceCell<T>];
-    _impl_init![ConstInit: <T: ConstInit> Self::new(T::INIT) =>
-        Cell<T>, ManuallyDrop<T>, RefCell<T>, UnsafeCell<T>
-    ];
-
-    _impl_init![ConstInit: <T: ConstInit> Self(T::INIT) =>
-        PanicAssertUnwindSafe<T>, Reverse<T>, Saturating<T>, Wrapping<T>
-    ];
-
-    _impl_init![ConstInit: <T> Self => PhantomData<T>]; // no need for trait bound here
+    // option
     _impl_init![ConstInit: <T: ConstInit> Some(T::INIT) => Option<T>];
-    impl<T: ConstInit, E> ConstInit for Result<T, E> {
-        const INIT: Self = { Ok(T::INIT) };
-    }
+
+    // panic
+    _impl_init![ConstInit: <T: ConstInit> Self(T::INIT) => PanicAssertUnwindSafe<T>];
+
+    // result
+    impl<T: ConstInit, E> ConstInit for Result<T, E> { const INIT: Self = { Ok(T::INIT) }; }
+
+    // time
+    _impl_init![ConstInit: Duration::new(0, 0) => Duration];
 
     // WAIT: [exclusive_wrapper](https://github.com/rust-lang/rust/issues/98407)
     // _impl_init![ConstInit: <T: ConstInit> Self::new(T::INIT) => Exclusive<T>];